+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Added
 - Add Go protobuf generation support ([#131](https://github.com/opensearch-project/opensearch-protobufs/pull/131))
 - Add protobuf zip generation for releases ([#139](https://github.com/opensearch-project/opensearch-protobufs/pull/139))
+- Add python wheel generation support ([#135](https://github.com/opensearch-project/opensearch-protobufs/pull/135))
 
 ### Changed
 - Remove json_name annotations and prefix fields with `underscore_`  ([#149](https://github.com/opensearch-project/opensearch-protobufs/pull/149))